fix linux bug ,version 0.2.9

+++ lib/ofxbundler/dsl.rb
@@ -4,6 +4,7 @@ require 'net/http'
 require 'zip/zip'
 module OfxBundler
     class Dsl
+        attr_accessor :_version
         @@latest_version = "007"
         @@configs={
             :osx=>{
@@ -23,7 +24,7 @@ module OfxBundler
                 },
                 "007_64" => {
                     "file"=> "http://www.openframeworks.cc/versions/preRelease_v0.07/of_preRelease_v007_linux64.tar.gz",
-                    "filename" => "of_preRelease_v007_linux64",
+                    "filename" => "of_preRelease_v007_linux64.tar.gz",
                     "dirname"=> "of_preRelease_v007_linux64"
                 }
 
@@ -31,7 +32,10 @@ module OfxBundler
         
         }
 
+
+
         def initialize
+            self._version = @@latest_version
         end
 
         def self.evalute(filename)
@@ -61,6 +65,7 @@ module OfxBundler
 
         #install latest version
         def ofx version=@@latest_version
+            self._version=version
             response = Faraday.get do |req| 
                 req.url "http://www.openframeworks.cc/download/"
             end
@@ -78,7 +83,7 @@ module OfxBundler
                     request = Net::HTTP::Get.new uri.request_uri
 
                     http.request request do |response|
-                        open 'ofx.zip', 'w' do |io|
+                        open config["filename"], 'w' do |io|
                             response.read_body do |chunk|
                                 io.write chunk
                             end
@@ -87,22 +92,28 @@ module OfxBundler
                 end
 
                 if File.exists?(config["dirname"])
-                   puts "openframewors dir exists. pass.." 
+                   puts "openframewors dir #{config['dirname']} exists. pass.." 
                 else
-                    puts "unzip dir.."
-                    destination="."
-                    Zip::ZipFile.open("ofx.zip") {|file|
-                    
-                        file.each do |f| 
-                            f_path = File.join(destination, f.name)
-                            FileUtils.mkdir_p(File.dirname(f_path))
-                            file.extract(f,f_path)
-                        end
-                    } 
-                    #cleanup
-                    p "cleanup..."
-                    `rm -rf __MACOSX`
-                    `rm -rf ofx.zip`
+
+                    if RUBY_PLATFORM.downcase.include?("darwin")
+                        puts "unzip dir.."
+                        destination="."
+                        Zip::ZipFile.open(config["filename"]) {|file|
+
+                            file.each do |f| 
+                                f_path = File.join(destination, f.name)
+                                FileUtils.mkdir_p(File.dirname(f_path))
+                                file.extract(f,f_path)
+                            end
+                        } 
+                        #cleanup
+                        p "cleanup..."
+                        `rm -rf __MACOSX`
+                        `rm -rf ofx.zip`
+                    elsif RUBY_PLATFORM.downcase.include?("linux")
+                        p "tar file... #{config['filename']}"
+                        `tar vzxf #{config["filename"]}`
+                    end
                 end
 
             end
@@ -112,14 +123,14 @@ module OfxBundler
 
         # install or update the addon
         def addon name,version=@@latest_version
-            config = get_config(version)
+            config = get_config(self._version)
             addon_author,addon_name = name.split("/")
-            if Dir.exists?("#{config["dirname"]}/addons/#{addon_name}")
+            if File.exists?("#{config["dirname"]}/addons/#{addon_name}")
                 puts "update #{name}"
                 `cd #{config["dirname"]}/addons/#{addon_name} && git pull` 
             else
                 puts "clone #{name}"
-               `cd #{config["dirname"]}/addons && git clone https://github.com/#{name}.git` 
+                `cd #{config["dirname"]}/addons && git clone https://github.com/#{name}.git` 
             end
         end
     end
